+++ changelog.md
@@ -1,1 +1,4 @@
-- changed required Majrusz Library version from 6.1.7+ to 7.0.0+
+- updated Chinese translation (thanks to @UDTakerBean)
+- made Misanthropy not obtainable via enchanting table on axes to match other vanilla enchantments
+- changed final Fishing Fanatic level text formatting
+- optimized Horse Frost Walker enchantment